Ignore type errors in `build/src/deps/`.
This might not be smart. I'm not sure what the repercussions of this
change are. I had an issue that the build was failing because of a type
failure in expect or assert. These dependencies are only used for tests
and not in the code. But if we add future dependencies which are used by
the code then we probably want those to be tested.